basic-straight-graph/pyproject.toml
[project]
name = "langgraph-tutorial"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "ipykernel>=6.29.5",
    "langchain[anthropic]>=0.3.24",
    "langgraph>=0.3.34",
    "python-dotenv>=1.1.0",
    "langchain_openai>=0.1.0"
]
